Update mqtt_client.dart
When an active connection to the MQTT server is lost. The _disconnect() method sets the authenticationManager to null. 
When you call connect() again on the client, you get an exception "Null check operator used on a null value". 
With this extra line of code, the authenticationManager has an object for sure.