Do not clear messageStream, was loosing messages.
Our situation is that we are sending messages pretty often and we were loosing
received messages. I could see that the package on the network and then traced
it into this part where messageStream actually contained more than one message
and all but the first was actually handled.

I moved the clear() into the else part as it seemed to help when reconnecting.